requirements.txt
pytesseract
discord.py
opencv-python